# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Company {
  id: string;
  cnpj?: string | null;
  name: string;
  phone?: string | null;
  whatsapp?: string | null;
  email?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  metaTitle?: string | null;
  metaDescription?: string | null;
  subdomain: string;
  htmlContent?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CompanyCreateInput {
  cnpj?: string | null;
  name: string;
  phone?: string | null;
  whatsapp?: string | null;
  email?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  metaTitle?: string | null;
  metaDescription?: string | null;
  subdomain: string;
}

export interface CompanyUpdateInput {
  cnpj?: string | null;
  name?: string;
  phone?: string | null;
  whatsapp?: string | null;
  email?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  metaTitle?: string | null;
  metaDescription?: string | null;
  subdomain?: string;
}

export interface DashboardStats {
  totalCompanies: number;
  companiesWithPages: number;
  companiesWithoutPages: number;
  statesRepresented: number;
}

export interface StateCount {
  state: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListCompaniesParams = {
  search?: string;
  state?: string;
  city?: string;
};

export type GetRecentCompaniesParams = {
  limit?: number;
};