# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SaaS Admin Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: companies
    description: Company management
  - name: dashboard
    description: Dashboard analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /companies/{id}/generate-page:
    post:
      operationId: generateCompanyPage
      tags: [companies]
      summary: Generate landing page HTML for a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: HTML generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  /companies/subdomain/{subdomain}:
    get:
      operationId: getCompanyBySubdomain
      tags: [companies]
      summary: Get company by subdomain
      parameters:
        - name: subdomain
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent:
    get:
      operationId: getRecentCompanies
      tags: [dashboard]
      summary: Get recently added companies
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Recent companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"

  /dashboard/by-state:
    get:
      operationId: getCompaniesByState
      tags: [dashboard]
      summary: Companies grouped by state
      responses:
        "200":
          description: Grouped counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StateCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Company:
      type: object
      properties:
        id:
          type: string
        cnpj:
          type: string
          nullable: true
        name:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        metaTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        subdomain:
          type: string
        htmlContent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, subdomain, createdAt, updatedAt]

    CompanyCreateInput:
      type: object
      properties:
        cnpj:
          type: string
          nullable: true
        name:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        metaTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        subdomain:
          type: string
      required: [name, subdomain]

    CompanyUpdateInput:
      type: object
      properties:
        cnpj:
          type: string
          nullable: true
        name:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        metaTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        subdomain:
          type: string

    DashboardStats:
      type: object
      properties:
        totalCompanies:
          type: integer
        companiesWithPages:
          type: integer
        companiesWithoutPages:
          type: integer
        statesRepresented:
          type: integer
      required: [totalCompanies, companiesWithPages, companiesWithoutPages, statesRepresented]

    StateCount:
      type: object
      properties:
        state:
          type: string
        count:
          type: integer
      required: [state, count]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]