# lib/api-zod/src/generated/types/companyCreateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyCreateInput {
  cnpj?: string | null;
  name: string;
  phone?: string | null;
  whatsapp?: string | null;
  email?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  metaTitle?: string | null;
  metaDescription?: string | null;
  subdomain: string;
}