# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalCompanies: number;
  companiesWithPages: number;
  companiesWithoutPages: number;
  statesRepresented: number;
}